% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userStream-package.R
\docType{package}
\name{userStream}
\alias{userStream}
\alias{userStream-package}
\title{userStream}
\description{
Stream Clustering algorithm applicable to custoemr segmentation. The algorithm allows to insert new but also update already processed observations.
}
\author{
Matthias Carnein \email{Matthias.Carnein@uni-muenster.de}
}
