% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userStream.R
\docType{class}
\name{userStream_R-class}
\alias{userStream_R-class}
\alias{userStream}
\title{Reference Class userStream}
\description{
Reference class mostly used to expose C class object
}
\section{Fields}{

\describe{
\item{\code{C}}{exposed C class}
}}


\author{
Matthias Carnein \email{matthias.carnein@uni-muenster.de}
}
